Keep Agent filter files under the oxlint line cap after rebase.
Move getDefaultWorkspaceSession and status-bar hydration out of files
that the filterAgentIds wiring pushed over 300 lines.
