+++ src/shared/default-workspace-session.ts
@@ -1,0 +1,20 @@
+import type { WorkspaceSessionState } from './workspace-session-state-types'
+
+export function getDefaultWorkspaceSession(): WorkspaceSessionState {
+  return {
+    activeRepoId: null,
+    activeWorktreeId: null,
+    activeTabId: null,
+    tabsByWorktree: {},
+    terminalLayoutsByTabId: {},
+    openFilesByWorktree: {},
+    markdownFrontmatterVisible: {},
+    browserTabsByWorktree: {},
+    browserPagesByWorkspace: {},
+    activeBrowserTabIdByWorktree: {},
+    activeFileIdByWorktree: {},
+    activeTabTypeByWorktree: {},
+    browserUrlHistory: [],
+    defaultTerminalTabsAppliedByWorktreeId: {}
+  }
+}
